app/requirements.txt
requests==2.32.3
selenium
python-dotenv==1.1.0